+++ app/src/core/operators/index.ts
@@ -1,3 +1,4 @@
+export * from './executiveCoordinator'
 export * from './operator'
 export * from './operatorEngine'
 export * from './operatorEvents'